+++ scalyr_agent/configuration.py
@@ -25,6 +25,7 @@ import os
 import re
 import socket
 import time
+import logging
 
 import six
 import six.moves.urllib.parse
@@ -851,6 +852,13 @@ class Configuration(object):
         return self.__get_config().get_int("debug_level")
 
     @property
+    def stdout_severity(self):
+        """Returns the configuration value for 'stdout_severity'.
+        Only used when running in no-fork mode.
+        """
+        return self.__get_config().get_string("stdout_severity")
+
+    @property
     def ca_cert_path(self):
         """Returns the configuration value for 'ca_cert_path'."""
         return self.__get_config().get_string("ca_cert_path")
@@ -1626,6 +1634,22 @@ class Configuration(object):
                 "The debug level must be between 0 and 5 inclusive",
                 "debug_level",
                 "badDebugLevel",
+            )
+
+        self.__verify_or_set_optional_string(
+            config,
+            "stdout_severity",
+            "NOTSET",
+            description,
+            apply_defaults,
+            env_aware=True,
+        )
+        stdout_severity = config.get_string("stdout_severity", default_value="NOTSET")
+        if not hasattr(logging, stdout_severity):
+            raise BadConfiguration(
+                "The stdout severity must be a valid logging level name",
+                "stdout_severity",
+                "badStdoutSeverity",
             )
 
         self.__verify_or_set_optional_float(